10行目double→intに変更。これにより13行目(int)(shina1+result)のint消せる。plintlnはあくまでoutputであり、計算結果の実態は10行目

+++ src/Review01.java
@@ -6,11 +6,11 @@ public class Review01 {
         int shina1 = 1500;
         double zeikin1 = 0.1;
 
-        //　double型resultを宣言、double型引値を２つ受け取るメソットの定義（戻り値なし）
-        double result = taxMethod1(shina1, zeikin1);
+        //　int型resultを宣言、double型引値を２つ受け取るメソットの定義（戻り値なし）
+        int result = taxMethod1(shina1, zeikin1);
 
         //　演算結果を表示する
-        System.out.println(shina1 + "円の商品の税込価格は" + (int)(shina1 + result) + "円(消費税は" + (int)result + "円)です。");
+        System.out.println(shina1 + "円の商品の税込価格は" + (shina1 + result) + "円(消費税は" + result + "円)です。");
     }
 
     // double型引数を２つ受け取るメソットの定義（戻り値あり）
